Updated front end react pages with updated driectory structure that was not refactored correctly IOU-17
